Fixed frames not being rendered after continueForTime

+++ build.sh
@@ -119,9 +119,11 @@ mkdir -p Unreal/Environments/Blocks/Plugins
 rsync -a --delete Unreal/Plugins/AirSim Unreal/Environments/Blocks/Plugins
 
 # Update Greenhouse project
-Unreal/Environments/Greenhouse/clean.sh
-mkdir -p Unreal/Environments/Greenhouse/Plugins
-rsync -a --delete Unreal/Plugins/AirSim Unreal/Environments/Greenhouse/Plugins
+if [ -d "Unreal/Environments/Greenhouse/clean.sh" ]; then
+    Unreal/Environments/Greenhouse/clean.sh
+    mkdir -p Unreal/Environments/Greenhouse/Plugins
+    rsync -a --delete Unreal/Plugins/AirSim Unreal/Environments/Greenhouse/Plugins
+fi
 
 set +x
